refactor: update path to Details component
