+++ src/components/Employment.tsx
@@ -2,7 +2,7 @@ import { useForm } from "react-hook-form";
 import { create } from "zustand";
 import { ChangeEvent } from "react";
 import { Summary } from "./Summary";
-import { Details } from "./Details";
+import { Details } from "./details/details";
 
 type State = {
   position: string;
